Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1718753 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: f0a818f4671f3988f48443f2414d4b6dd8dc0748

+++ src/core/org/apache/jmeter/report/processor/StatisticsSummaryConsumer.java
@@ -55,9 +55,9 @@ public class StatisticsSummaryConsumer extends
 	}
 
 	long elapsedTime = sample.getElapsedTime();
-	data.getPercentile1().addValue((double) elapsedTime);
-	data.getPercentile2().addValue((double) elapsedTime);
-	data.getPercentile3().addValue((double) elapsedTime);
+	data.getPercentile1().addValue(elapsedTime);
+	data.getPercentile2().addValue(elapsedTime);
+	data.getPercentile3().addValue(elapsedTime);
 
 	data.setMin(elapsedTime);
 	data.setMax(elapsedTime);
